% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_means.R
\name{multiple_means}
\alias{multiple_means}
\title{Table showing mean, SD, range, for a multiple numeric variables}
\usage{
multiple_means(data, group, caption = NULL, table = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{group}{The grouping variable, e.g. trial arm}

\item{caption}{A caption for the table (string)}

\item{table}{Optional parameter to determine whether the function
should return a formatted \code{pander} or \code{flextable} table
(\code{TRUE}) or the raw data frame (\code{FALSE}). Defaults to \code{TRUE}.}

\item{...}{Comma-separated list of variables to calculate means for.}
}
\value{
A data frame with descriptive statistecs (mean, SD, range)
    the selected variables; a formatted table with this information,
    depending on the \code{table} option.
}
\description{
Table showing mean, SD, range, for a multiple numeric variables
}
\examples{
    data(mtcars)
    multiple_means(mtcars,
                   group = cyl,
                   caption = "My table",
                   table = FALSE,
                   mpg, disp, hp)
}
