% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_counts.R
\name{table_of_counts}
\alias{table_of_counts}
\title{Table showing counts for one or more categorical variables.}
\usage{
table_of_counts(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Comma-separated list of categorical variables to include.}
}
\value{
A \code{flextable} with descriptive statistics (N, percent, range)
    for the selected variables.
}
\description{
Table showing counts for one or more categorical variables.
}
\examples{
    data(mtcars)
    table_of_counts(mtcars, cyl, wt, hp)

}
