% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_means.R
\name{table_of_means}
\alias{table_of_means}
\title{Table showing mean, SD, range, for one or more numeric variables}
\usage{
table_of_means(data, group, ...)
}
\arguments{
\item{data}{A data frame}

\item{group}{The grouping variable, e.g. trial arm}

\item{...}{Comma-separated list of variables to calculate means for.}
}
\value{
A data frame with descriptive statistics (mean, SD, range)
    for the selected variables.
}
\description{
Table showing mean, SD, range, for one or more numeric variables
}
\examples{
    data(mtcars)
    table_of_means(mtcars,
                   cyl,
                   wt, hp)

    data(starwars)
    table_of_means(starwars,
                   gender,
                   height, mass, birth_year)
}
