% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_medians.R
\name{table_of_medians}
\alias{table_of_medians}
\title{Table showing median, IQR, and range, for one or more numeric variables}
\usage{
table_of_medians(data, group, ...)
}
\arguments{
\item{data}{A data frame}

\item{group}{The grouping variable, e.g. trial arm}

\item{...}{Comma-separated list of variables to calculate medians for.}
}
\value{
A data frame with descriptive statistics (median, IQR, range)
    for the selected variables.
}
\description{
Table showing median, IQR, and range, for one or more numeric variables
}
\examples{
    data(mtcars)
    table_of_medians(mtcars,
                     group = cyl,
                     wt, hp)

    data(starwars)
    table_of_means(starwars,
                   group = gender,
                   height, mass, birth_year)
}
