% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_counts.R, R/table_of_counts_improved.R
\name{table_of_counts}
\alias{table_of_counts}
\alias{table_of_counts}
\title{Table showing counts for one or more categorical variables.}
\usage{
table_of_counts(data, group)

table_of_counts(data, group)
}
\arguments{
\item{data}{A data frame}

\item{group}{An optional grouping variable.}

\item{data}{A data frame}

\item{group}{An optional grouping variable.}
}
\value{
A \code{flextable} with descriptive statistics (N, percent, range)
    for all variables in the supplied \code{data.frame}.

A \code{flextable} with descriptive statistics (N, percent, range)
    for all variables in the supplied \code{data.frame}.
}
\description{
Table showing counts for one or more categorical variables.

Table showing counts for one or more categorical variables.
}
\examples{
    data(mtcars)
    mtcars \%>\%
      select(cyl, vs) \%>\%
      table_of_counts()
      
    mtcars \%>\%
      select(cyl, vs, am) \%>\%
      table_of_counts(cyl)

    data(mtcars)
    mtcars \%>\%
      select(cyl, vs) \%>\%
      table_of_counts()
      
    mtcars \%>\%
      select(cyl, vs, am) \%>\%
      table_of_counts(cyl)

}
