% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_means.R
\name{table_of_means}
\alias{table_of_means}
\title{Table showing mean, SD, range, for one or more numeric variables}
\usage{
table_of_means(data, ..., group)
}
\arguments{
\item{data}{A data frame}

\item{...}{Comma-separated list of variables to calculate means for.}

\item{group}{An optional grouping variable, e.g. trial arm}
}
\value{
A data frame with descriptive statistics (mean, SD, range)
    for the selected variables.
}
\description{
Table showing mean, SD, range, for one or more numeric variables
}
\examples{
    data(mtcars)
    table_of_means(mtcars,
                   wt, hp,
                   group = cyl)

    data(starwars)
    table_of_means(starwars, height, mass, birth_year)
    table_of_means(starwars, height, mass, birth_year,
                   group = gender)
    
}
